add clouflare and react router
