+++ app/routes/play/components/header/header.tsx
@@ -1,0 +1,44 @@
+import { IconButton } from "../../components/button/icon-button";
+import { useTimer } from "../../hooks/use-timer";
+
+export function Header() {
+  const timer = useTimer();
+
+  return (
+    <header className="sticky inset-0 top-0 header py-1 backdrop-blur-sm backdrop-opacity-95 border rounded-md flex items-center justify-between px-2">
+      <IconButton title="Back">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          fill="none"
+          viewBox="0 0 24 24"
+          strokeWidth={1.5}
+          stroke="currentColor"
+          className="size-6"
+        >
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            d="M15.75 19.5 8.25 12l7.5-7.5"
+          />
+        </svg>
+      </IconButton>
+      <p className="text-sm tabular-nums text-pretty">{timer}</p>
+      <IconButton title="Back">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          fill="none"
+          viewBox="0 0 24 24"
+          strokeWidth={1.5}
+          stroke="currentColor"
+          className="size-6"
+        >
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            d="M6.75 12a.75.75 0 1 1-1.5 0 .75.75 0 0 1 1.5 0ZM12.75 12a.75.75 0 1 1-1.5 0 .75.75 0 0 1 1.5 0ZM18.75 12a.75.75 0 1 1-1.5 0 .75.75 0 0 1 1.5 0Z"
+          />
+        </svg>
+      </IconButton>
+    </header>
+  );
+}
