+++ package.json
@@ -1,22 +1,35 @@
 {
-  "name": "spell-chain",
+  "name": "word-combo",
   "private": true,
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
+    "old:dev": "vite",
+    "old:build": "tsc && vite build",
     "lint": "eslint src --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "build": "react-router build",
+    "dev": "react-router dev",
+    "start": "wrangler dev",
+    "deploy": "wrangler deploy",
+    "typecheck": "react-router typegen && tsc"
   },
   "dependencies": {
+    "@react-router/node": "^7.0.2",
+    "@react-router/serve": "^7.0.2",
     "clsx": "^2.1.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "isbot": "^5.1.17",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router": "^7.0.2",
     "sonner": "^1.7.0",
     "tailwind-merge": "^2.5.5"
   },
   "devDependencies": {
+    "@cloudflare/workers-types": "^4.20241202.0",
+    "@hiogawa/vite-node-miniflare": "^0.1.1",
+    "@react-router/dev": "^7.0.2",
+    "@types/node": "^22.10.1",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "@typescript-eslint/eslint-plugin": "^5.57.1",
@@ -26,10 +39,11 @@
     "eslint": "^8.38.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.3.4",
-    "postcss": "^8.4.23",
-    "tailwindcss": "^3.3.2",
-    "typescript": "^5.0.2",
-    "vite": "^4.3.2",
-    "vite-tsconfig-paths": "^5.0.1"
+    "postcss": "^8.4.49",
+    "tailwindcss": "^3.4.15",
+    "typescript": "^5.6.3",
+    "vite": "^5.4.11",
+    "vite-tsconfig-paths": "^5.0.1",
+    "wrangler": "^3.92.0"
   }
 }
