--- src/components/game-provider/types.ts
@@ -1,15 +1,0 @@
-export interface GameState {
-  count: number;
-  board: Map<string, string[]>;
-  currentWord: string[];
-  wordsOfTheDay: {
-    start: string;
-    target: string;
-  };
-  message: string;
-}
-
-export interface GameActions {
-  type: "message" | "clicked_key";
-  payload?: string;
-}
