+++ vite.config.ts
@@ -1,8 +1,49 @@
+import { vitePluginViteNodeMiniflare } from "@hiogawa/vite-node-miniflare";
+import { reactRouter } from "@react-router/dev/vite";
+import autoprefixer from "autoprefixer";
+import tailwindcss from "tailwindcss";
 import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react-swc";
 import tsconfigPaths from "vite-tsconfig-paths";
 
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react(), tsconfigPaths()],
-});
+export default defineConfig(({ isSsrBuild }) => ({
+  build: {
+    rollupOptions: isSsrBuild
+      ? {
+          input: "./workers/app.ts",
+        }
+      : undefined,
+  },
+  css: {
+    postcss: {
+      plugins: [tailwindcss, autoprefixer],
+    },
+  },
+  ssr: {
+    target: "webworker",
+    noExternal: true,
+    resolve: {
+      conditions: ["workerd", "browser"],
+    },
+    optimizeDeps: {
+      include: [
+        "react",
+        "react/jsx-runtime",
+        "react/jsx-dev-runtime",
+        "react-dom",
+        "react-dom/server",
+        "react-router",
+      ],
+    },
+  },
+  plugins: [
+    vitePluginViteNodeMiniflare({
+      entry: "./workers/app.ts",
+      miniflareOptions: (options) => {
+        options.compatibilityDate = "2024-11-18";
+        options.compatibilityFlags = ["nodejs_compat"];
+      },
+    }),
+    reactRouter(),
+    tsconfigPaths(),
+  ],
+}));
